refactor #47 : eventlistener 및 event 위치 이동
