+++ src/main/java/com/irum/orderservice/domain/coupon/event/CouponValidatedEvent.java
@@ -1,4 +1,4 @@
-package com.irum.orderservice.domain.order.event;
+package com.irum.orderservice.domain.coupon.event;
 
 import java.util.List;
 import java.util.UUID;
